refactor(lpa-questionnaire-web-app): rename interested-parties to representations
