+++ packages/lpa-questionnaire-web-app/src/lib/views.js
@@ -4,7 +4,7 @@ const VIEW = {
   OTHER_APPEALS: 'other-appeals',
   ACCURACY_SUBMISSION: 'accuracy-submission',
   EXTRA_CONDITIONS: 'extra-conditions',
-  INTERESTED_PARTIES: 'interested-parties',
+  REPRESENTATIONS: 'representations',
   NOTIFYING_PARTIES: 'notifying-parties',
   DEVELOPMENT_PLAN: 'development-plan',
   UPLOAD_PLANS: 'upload-plans',
