+++ packages/lpa-questionnaire-web-app/src/routes/index.js
@@ -9,7 +9,7 @@ const confirmAnswersRouter = require('./confirm-answers');
 const accuracySubmissionRouter = require('./accuracy-submission');
 const otherAppealsRouter = require('./other-appeals');
 const extraConditionsRouter = require('./extra-conditions');
-const { router: interestedPartiesRouter } = require('./interested-parties');
+const { router: representationsRouter } = require('./representations');
 const { router: notifyingPartiesRouter } = require('./notifying-parties');
 const developmentPlanRouter = require('./development-plan');
 const { router: uploadPlansRouter } = require('./upload-plans');
@@ -25,7 +25,7 @@ router.use(confirmAnswersRouter);
 router.use(accuracySubmissionRouter);
 router.use(otherAppealsRouter);
 router.use(extraConditionsRouter);
-router.use(interestedPartiesRouter);
+router.use(representationsRouter);
 router.use(notifyingPartiesRouter);
 router.use(developmentPlanRouter);
 router.use(uploadPlansRouter);
