+++ packages/lpa-questionnaire-web-app/tests/unit/routes/index.test.js
@@ -7,7 +7,7 @@ const confirmAnswersRouter = require('../../../src/routes/confirm-answers');
 const accuracySubmissionRouter = require('../../../src/routes/accuracy-submission');
 const otherAppealsRouter = require('../../../src/routes/other-appeals');
 const extraConditionsRouter = require('../../../src/routes/extra-conditions');
-const { router: interestedPartiesRouter } = require('../../../src/routes/interested-parties');
+const { router: interestedPartiesRouter } = require('../../../src/routes/representations');
 const { router: notifyingPartiesRouter } = require('../../../src/routes/notifying-parties');
 const developmentPlanRouter = require('../../../src/routes/development-plan');
 const { router: uploadPlansRouter } = require('../../../src/routes/upload-plans');
